Create migration for gamesessionmembers

+++ wavelength-api/Persistence/Repositories/GameSessionRepository.cs
@@ -8,7 +8,6 @@ namespace Persistence.Repositories;
 
 public class GameSessionRepository : IGameSessionRepository
 {
-
     private readonly DataContext context;
     private readonly IMapper mapper;
 
@@ -17,15 +16,15 @@ public class GameSessionRepository : IGameSessionRepository
         this.context = context;
         this.mapper = mapper;
     }
-    
+
     public async Task<GameSessionDTO> Create(Guid ownerID)
     {
         var newGameSession = new GameSession
         {
             OwnerID = ownerID,
-            JoinCode = GenerateRandomJoinCode(),
+            JoinCode = GenerateRandomJoinCode()
         };
-        
+
         context.GameSessions.Add(newGameSession);
 
         // Have the user join the game
@@ -47,9 +46,9 @@ public class GameSessionRepository : IGameSessionRepository
     {
         var gameSessionMember = new GameSessionMember
         {
-            UserID = userID,
+            UserID = userID.ToString(),
             GameSessionID = gameSessionID,
-            Team = 0,
+            Team = 0
         };
 
         context.GameSessionMembers.Add(gameSessionMember);
@@ -66,45 +65,15 @@ public class GameSessionRepository : IGameSessionRepository
             .FirstOrDefaultAsync();
         var gameSessionMember = await context.GameSessionMembers
             .Where(gsm => gsm.GameSessionID == gameSessionID)
-            .Where(gsm => gsm.UserID == userID)
+            .Where(gsm => gsm.UserID == userID.ToString())
             .FirstOrDefaultAsync();
 
         // If the game sessions owner is the user, delete the game session
-        if (gameSession != null && gameSession.OwnerID == userID)
-        {
-            context.GameSessions.Remove(gameSession);
-        }
+        if (gameSession != null && gameSession.OwnerID == userID) context.GameSessions.Remove(gameSession);
 
-        if (gameSessionMember != null)
-        {
-            context.GameSessionMembers.Remove(gameSessionMember);
-        }
+        if (gameSessionMember != null) context.GameSessionMembers.Remove(gameSessionMember);
 
         return await context.SaveChangesAsync() > 0;
-    }
-    
-    private String GenerateRandomJoinCode()
-    {
-        Random random = new Random();
-        String joinCode = "";
-
-        while (joinCode.Length == 0)
-        {
-            /// generate a value between 000,000 and 999,999 inclusive
-            int randomCode = random.Next(1000000);
-            String formattedCode = randomCode.ToString("D6");
-
-            var isCodeInUse = context.GameSessions
-                .Where(gs => gs.EndTime == null)
-                .Any(gs => gs.JoinCode == formattedCode);
-
-            if (!isCodeInUse)
-            {
-                joinCode = formattedCode;
-            }
-        }
-
-        return joinCode;
     }
 
     public async Task<bool> End(Guid gameSessionID)
@@ -112,7 +81,7 @@ public class GameSessionRepository : IGameSessionRepository
         var gameSession = await context.GameSessions
             .Where(gs => gs.ID == gameSessionID)
             .FirstOrDefaultAsync();
-        
+
         gameSession.EndTime = DateTime.Now;
 
         return await context.SaveChangesAsync() > 0;
@@ -123,9 +92,30 @@ public class GameSessionRepository : IGameSessionRepository
         var gameSession = await context.GameSessions
             .Where(gs => gs.ID == gameSessionID)
             .FirstOrDefaultAsync();
-        
+
         gameSession.StartTime = DateTime.Now;
 
         return await context.SaveChangesAsync() > 0;
+    }
+
+    private string GenerateRandomJoinCode()
+    {
+        var random = new Random();
+        var joinCode = "";
+
+        while (joinCode.Length == 0)
+        {
+            /// generate a value between 000,000 and 999,999 inclusive
+            var randomCode = random.Next(1000000);
+            var formattedCode = randomCode.ToString("D6");
+
+            var isCodeInUse = context.GameSessions
+                .Where(gs => gs.EndTime == null)
+                .Any(gs => gs.JoinCode == formattedCode);
+
+            if (!isCodeInUse) joinCode = formattedCode;
+        }
+
+        return joinCode;
     }
 }
